requirements.txt
blinker==1.8.2
boltons==24.0.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-HTTPAuth==4.8.0
Flask-CORS==6.0.0
google-auth==2.30.0
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
kubernetes==30.1.0
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==24.1
pyasn1==0.6.0
pyasn1-modules==0.4.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
six==1.16.0
urllib3==1.26.19
websocket-client==1.8.0
Werkzeug==3.0.3
SQLAlchemy==2.0.36
alembic==1.13.2
psycopg2-binary==2.9.9
redis==5.0.8
Flask-Limiter==3.8.0
PyJWT==2.9.0
bcrypt==4.2.1
